Add: support for deserializing multiple instances of a type when present in TTL source
This PR contains code that adds support for the deserializer to instansiate multiple distinct type instances when present in a source ttl.
The interface is slightly modified to support returning either a type T when only a single instance is present in the source or an array of type T when multiple.